backend/requirements.txt
asgiref==3.7.2
boto3==1.34.5
botocore==1.34.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
future==0.18.3
geocoder==1.38.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
ratelim==0.1.6
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
whitenoise==6.6.0
